made a basic webpage for caption generation
Another webpage called temp was created.

This webpage features basic caption generation and image uploading, similar to the image generation page.

To test out the new functionalities navigate to /generatechoose or /temp.

On the animals tab, the play button should lead to the temp generation page and the back button in the temp page should lead back to generate choose.

The generation page is still incomplete though, we are still trying to implement the report function ,multiple model category support and save functionality.