adjusted model selection page
The buttons of the model selection page have been coded to be able to pass in values into the /temp generate page.

For now, this only changes the color,title and image of the temp generate page. The code can be adjusted to include model type as a variable to be passed in.

to test the feature, go to /generateChoose and click the buttons. They all link to the /temp  generate page